move map components into subdirectory
