+++ src/pages/map/components/CategoryPicker.js
