+++ src/pages/map/components/Detail.test.js
