+++ src/pages/map/components/Pin.js
