+++ src/pages/map/components/SubmitTipFooter.js
